feat: add crud for memories

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+  // buscar todas as memórias
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,8 +25,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
       };
     });
   });
-
-  app.get("/memories/:id", async (request) => {
+  // buscar memória específica
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -31,6 +38,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+    // Se a memória não for pública e o id do usuário dono da memória for diferente do id do logado
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     return memory;
   });
@@ -49,14 +60,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "5ac8562f-5c67-49da-8319-5646cf8b8b6d",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
-
-  app.put("/memories/:id", async (request) => {
+  // edição de memória
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -70,8 +81,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+    // preciso verificar se o user que vai editar é o dono da memória
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
 
-    const memory = await prisma.memory.update({
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -84,13 +105,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     return memory;
   });
-
-  app.delete("/memories/:id", async (request) => {
+  // apagar memória
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+    // verificar se o usuário que deseja apagar a memória é dono dela
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
